More robust conversion of numeric literals
