+++ packages/typegpu/src/data/wgslTypes.ts
@@ -864,6 +864,7 @@ export interface Vec2f extends
     & ((v: AnyNumericVec2Instance) => v2f)
   > {
   readonly type: 'vec2f';
+  readonly primitive: F32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v2f;
@@ -884,6 +885,7 @@ export interface Vec2h extends
     & ((v: AnyNumericVec2Instance) => v2h)
   > {
   readonly type: 'vec2h';
+  readonly primitive: F16;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v2h;
@@ -904,6 +906,7 @@ export interface Vec2i extends
     & ((v: AnyNumericVec2Instance) => v2i)
   > {
   readonly type: 'vec2i';
+  readonly primitive: I32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v2i;
@@ -924,6 +927,7 @@ export interface Vec2u extends
     & ((v: AnyNumericVec2Instance) => v2u)
   > {
   readonly type: 'vec2u';
+  readonly primitive: U32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v2u;
@@ -945,6 +949,7 @@ export interface Vec2b extends
     & ((v: v2b) => v2b)
   > {
   readonly type: 'vec2<bool>';
+  readonly primitive: Bool;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v2b;
@@ -966,6 +971,7 @@ export interface Vec3f extends
     & ((x: number, v0: AnyNumericVec2Instance) => v3f)
   > {
   readonly type: 'vec3f';
+  readonly primitive: F32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v3f;
@@ -988,6 +994,7 @@ export interface Vec3h extends
     & ((x: number, v0: AnyNumericVec2Instance) => v3h)
   > {
   readonly type: 'vec3h';
+  readonly primitive: F16;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v3h;
@@ -1010,6 +1017,7 @@ export interface Vec3i extends
     & ((x: number, v0: AnyNumericVec2Instance) => v3i)
   > {
   readonly type: 'vec3i';
+  readonly primitive: I32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v3i;
@@ -1032,6 +1040,7 @@ export interface Vec3u extends
     & ((x: number, v0: AnyNumericVec2Instance) => v3u)
   > {
   readonly type: 'vec3u';
+  readonly primitive: U32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v3u;
@@ -1055,6 +1064,7 @@ export interface Vec3b extends
     & ((x: boolean, v0: v2b) => v3b)
   > {
   readonly type: 'vec3<bool>';
+  readonly primitive: Bool;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v3b;
@@ -1080,6 +1090,7 @@ export interface Vec4f extends
     & ((x: number, y: number, v0: AnyNumericVec2Instance) => v4f)
   > {
   readonly type: 'vec4f';
+  readonly primitive: F32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v4f;
@@ -1106,6 +1117,7 @@ export interface Vec4h extends
     & ((x: number, y: number, v0: AnyNumericVec2Instance) => v4h)
   > {
   readonly type: 'vec4h';
+  readonly primitive: F16;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v4h;
@@ -1132,6 +1144,7 @@ export interface Vec4i extends
     & ((x: number, y: number, v0: AnyNumericVec2Instance) => v4i)
   > {
   readonly type: 'vec4i';
+  readonly primitive: I32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v4i;
@@ -1158,6 +1171,7 @@ export interface Vec4u extends
     & ((x: number, y: number, v0: AnyNumericVec2Instance) => v4u)
   > {
   readonly type: 'vec4u';
+  readonly primitive: U32;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v4u;
@@ -1185,6 +1199,7 @@ export interface Vec4b extends
     & ((x: boolean, y: boolean, v0: v2b) => v4b)
   > {
   readonly type: 'vec4<bool>';
+  readonly primitive: Bool;
 
   // Type-tokens, not available at runtime
   readonly [$repr]: v4b;
