+++ packages/typegpu/src/std/operators.ts
@@ -13,16 +13,15 @@ import {
   type mBaseForVec,
   type vBaseForMat,
 } from '../data/wgslTypes.ts';
-import { convertToCommonType } from '../tgsl/generationHelpers.ts';
+import { convertToCommonType } from '../tgsl/conversion.ts';
 import { getResolutionCtx } from '../execMode.ts';
 import type { ResolutionCtx } from '../types.ts';
 import { $internal } from '../shared/symbols.ts';
-import { f16, f32 } from '../data/numeric.ts';
+import { abstractFloat, f16, f32 } from '../data/numeric.ts';
 
 function tryUnify<T extends Snippet[]>(
   values: T,
   restrictTo?: AnyWgslData[],
-  concretizeTypes = false,
   verbose = true,
 ): T {
   const ctx = getResolutionCtx() as ResolutionCtx;
@@ -30,7 +29,6 @@ function tryUnify<T extends Snippet[]>(
     ctx,
     values,
     restrictTo,
-    concretizeTypes,
     verbose,
   });
   return converted ?? values;
@@ -247,8 +245,20 @@ export const div = createDualImpl(
   // CPU implementation
   cpuDiv,
   // CODEGEN implementation
-  (lhs, rhs) => {
-    const [convLhs, convRhs] = tryUnify([lhs, rhs], [f32, f16], true, false);
+  (...args) => {
+    // const resultType = isSnippetNumeric(convLhs)
+    //   ? convRhs.dataType
+    //   : convLhs.dataType;
+
+    const [lhs, rhs] = tryUnify(args, [f32, f16], false);
+    let convLhs = lhs;
+    let convRhs = rhs;
+    if (convLhs.dataType.type === 'abstractInt') {
+      convLhs = snip(convLhs.value, abstractFloat);
+    }
+    if (convRhs.dataType.type === 'abstractInt') {
+      convRhs = snip(convRhs.value, abstractFloat);
+    }
 
     if (
       (typeof lhs.value === 'number' || isVecInstance(lhs.value)) &&
@@ -257,11 +267,11 @@ export const div = createDualImpl(
       // Precomputing
       return snip(
         cpuDiv(lhs.value as never, rhs.value as never),
-        convLhs.dataType,
+        lhs.dataType,
       );
     }
 
-    return snip(stitch`(${convLhs} / ${convRhs})`, convLhs.dataType);
+    return snip(stitch`(${lhs} / ${rhs})`, lhs.dataType);
   },
   'div',
 );
